fix(audio): promote int operands before they meet a double in frame and Bessel math
AudioFormat.BytesPerFrame multiplied SampleRate by BytesPerSample in int
arithmetic before scaling by the frame duration, so a rate/width product
past int.MaxValue wrapped negative: 1 GHz at 4 bytes over 100 us gave
-29496 bytes instead of 400000. ResamplerCoefficients.BesselI0 divided by
k * k in int as well; k stops at 30 there, so it could not overflow, but
it now divides by a double too.

Promoting one operand leaves every real format bit-identical. A before and
after dump of BytesPerFrame and SamplesPerFrame for eight formats over
2008 durations, BesselI0 over 0..20 and all twelve coefficient tables
differed only in the overflow case. The new tests pin representative
frame sizes, BesselI0 bit patterns and three coefficient tables; BesselI0
becomes internal and the test assembly gains InternalsVisibleTo to reach it.

Resolves CodeQL cs/loss-of-precision in AudioFormat.cs and ResamplerCoefficients.cs.

+++ src/Verbara.Sdk.Audio/AudioFormat.cs
@@ -31,7 +31,8 @@ public readonly record struct AudioFormat(
     public int BytesPerFrame(TimeSpan frameDuration)
     {
         ArgumentOutOfRangeException.ThrowIfLessThanOrEqual(frameDuration, TimeSpan.Zero, nameof(frameDuration));
-        return (int)(SampleRate * BytesPerSample * frameDuration.TotalSeconds);
+        // Promote before multiplying: rate x bytes-per-sample must not wrap in int arithmetic.
+        return (int)((double)SampleRate * BytesPerSample * frameDuration.TotalSeconds);
     }
 
     /// <summary>Number of samples per frame of the given duration.</summary>
